+++ src/lib/punctuation/dash.js
@@ -114,7 +114,7 @@ export function fixDashBetweenCardinalNumbers(string, locale) {
 	@returns {string} — output with en dash between percentage range
 */
 export function fixDashBetweenPercentageRange(string, locale) {
-	let pattern = "(%)([" + locale.spaces + "]?[" + locale.hyphen + locale.enDash + locale.emDash + "][" + locale.spaces + "]?)(" + locale.cardinalNumber + ")";
+	let pattern = "([" + locale.percent + locale.permille + locale.permyriad + "])([" + locale.spaces + "]?[" + locale.hyphen + locale.enDash + locale.emDash + "][" + locale.spaces + "]?)(" + locale.cardinalNumber + ")";
 	let re = new RegExp(pattern, "g");
 	let replacement = "$1" + locale.enDash + "$3";
 	return string.replace(re, replacement);