+++ src/lib/whitespace/nbsp.js
@@ -88,6 +88,23 @@ export function addNbspAfterRomanNumeral(string, locale) {
 }
 
 
+
+/*
+	Fix nbsp before % (percent), ‰ (permille) and ‱ (permyriad)
+
+	@param {string} string — input text for identification
+	@returns {string} — output with correctly added non-breaking space
+*/
+export function addNbspBeforePercent(string, locale) {
+	let pattern = "([" + locale.spaces + "])([" + locale.percent + locale.permille + locale.permyriad + "])";
+	let re = new RegExp(pattern, "g");
+	let replacement = locale.nbsp + "$2";
+
+	return string.replace(re, replacement);
+}
+
+
+
 /*
 	Helper function that adds nbsp after symbols
 	in their respective *.js files
@@ -136,6 +153,7 @@ export function fixNbsp(string, locale) {
 	string = addNbspAfterOrdinalNumber(string, locale);
 	string = addNbspWithinOrdinalDate(string, locale);
 	string = addNbspAfterRomanNumeral(string, locale);
+	string = addNbspBeforePercent(string, locale);
 
 	return string;
 }